+++ OpenRTM_aist_v121e_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/InPortPahoSubJson.py
@@ -4,7 +4,7 @@
 ##
 # @file   InPortPahoSubJson.py
 # @brief  InPortPahoSubJson class
-# @date   2020/12/07
+# @date   2020/12/09
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -18,18 +18,11 @@ from omniORB import any
 import OpenRTM_aist
 import OpenRTM__POA,OpenRTM
 import RTC
-import signal
-import os
 import time
-import threading
 import sys
 from OpenRTM_aist_paho_mqtt_module.paho_client.PahoSubscriber import PahoSubscriber
 from OpenRTM_aist_paho_mqtt_module.reserializer.DataTypeFormat import DataTypeFormat
-
-# There was a Ctrl+C interruption or not
-stop = False
-# Constructor was called already or not
-called = False
+from OpenRTM_aist.ManagerActionListener import ManagerActionListeners
 
 ##
 # @class InPortPahoSubJson
@@ -41,30 +34,9 @@ class InPortPahoSubJson(OpenRTM_aist.InPortProvider, PahoSubscriber):
   """
 
   ##
-  # @brief Signal handler
-  #
-  @staticmethod
-  def signal_handler(num, frame):
-    global stop
-    print(" Ctrl+C interrupted.")
-    stop = True
-    if called == False:
-      os._exit(0)
-
-  ##
-  # @brief Shutdown hook
-  #
-  def catch_signal(self):
-    while not stop:
-      time.sleep(1)
-    self.__del__()
-    os._exit(0)
-
-  ##
   # @brief Constructor
   #
   def __init__(self):
-    global called
     OpenRTM_aist.InPortProvider.__init__(self)
     PahoSubscriber.__init__(self)
 
@@ -74,15 +46,11 @@ class InPortPahoSubJson(OpenRTM_aist.InPortProvider, PahoSubscriber):
     self._profile = None
     self._listeners = None
 
-    orb = OpenRTM_aist.Manager.instance().getORB()
-
     callback = self.on_message
     PahoSubscriber.set_on_message(self, callback)
 
-    thread = threading.Thread(target=self.catch_signal)
-    thread.daemon = True
-    thread.start()
-    called = True
+    self._mgr = OpenRTM_aist.Manager.instance()
+    self._mgr.addManagerActionListener(ManagerActionListener(self))
 
     return
 
@@ -135,22 +103,18 @@ class InPortPahoSubJson(OpenRTM_aist.InPortProvider, PahoSubscriber):
       cdrmsg = self.__formatter.reserializeFromJsonToCdr(data)
 
       if not self._buffer:
-        #self.onReceiverError(data)
         self.onReceiverError(cdrmsg)
         return OpenRTM.PORT_ERROR
 
       self._rtcout.RTC_PARANOID("received data size: %d", len(data))
 
-      #self.onReceived(data)
       self.onReceived(cdrmsg)
 
       if not self._connector:
         return OpenRTM.PORT_ERROR
 
-      #ret = self._connector.write(data)
       ret = self._connector.write(cdrmsg)
 
-      #return self.convertReturn(ret, data)
       return self.convertReturn(ret, cdrmsg)
 
     except:
@@ -461,9 +425,27 @@ class InPortPahoSubJson(OpenRTM_aist.InPortProvider, PahoSubscriber):
     self.__datatype = OpenRTM_aist.instantiateDataType(eval(tmp_datatype))
 
 ##
-# @brief Catch ctrl+c interruption
+# @class ManagerActionListener
+# @brief ManagerActionListener class
 #
-signal.signal(signal.SIGINT, InPortPahoSubJson.signal_handler)
+class ManagerActionListener(ManagerActionListeners):
+  def __init__(self, InPortPahoSubJson):
+    self._InPortPahoSubJson = InPortPahoSubJson
+
+  def preShutdown(self):
+    pass
+
+  ##
+  # @brief Clean up mqtt communication module instance when RTC exit
+  #
+  def postShutdown(self):
+    self._InPortPahoSubJson.__del__()
+
+  def preReinit(self):
+    pass
+
+  def postReinit(self):
+    pass
 
 ##
 # @brief Initialize InPortPahoSubJson module