+++ OpenRTM_aist_v121e_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/InPortPahoSubscriber.py
@@ -4,7 +4,7 @@
 ##
 # @file   InPortPahoSubscriber.py
 # @brief  InPortPahoSubscriber class
-# @date   2020/11/20
+# @date   2020/12/09
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -17,17 +17,10 @@
 from omniORB import any
 import OpenRTM_aist
 import OpenRTM__POA,OpenRTM
-import signal
-import os
 import time
-import threading
 import sys
 from OpenRTM_aist_paho_mqtt_module.paho_client.PahoSubscriber import PahoSubscriber
-
-# There was a Ctrl+C interruption or not
-stop = False
-# Constructor was called already or not
-called = False
+from OpenRTM_aist.ManagerActionListener import ManagerActionListeners
 
 ##
 # @class InPortPahoSubscriber
@@ -39,30 +32,9 @@ class InPortPahoSubscriber(OpenRTM_aist.InPortProvider, PahoSubscriber):
   """
 
   ##
-  # @brief Signal handler
-  #
-  @staticmethod
-  def signal_handler(num, frame):
-    global stop
-    print(" Ctrl+C interrupted.")
-    stop = True
-    if called == False:
-      os._exit(0)
-
-  ##
-  # @brief Shutdown hook
-  #
-  def catch_signal(self):
-    while not stop:
-      time.sleep(1)
-    self.__del__()
-    os._exit(0)
-
-  ##
   # @brief Constructor
   #
   def __init__(self):
-    global called
     OpenRTM_aist.InPortProvider.__init__(self)
     PahoSubscriber.__init__(self)
 
@@ -72,15 +44,11 @@ class InPortPahoSubscriber(OpenRTM_aist.InPortProvider, PahoSubscriber):
     self._profile = None
     self._listeners = None
 
-    orb = OpenRTM_aist.Manager.instance().getORB()
-
     callback = self.on_message
     PahoSubscriber.set_on_message(self, callback)
 
-    thread = threading.Thread(target=self.catch_signal)
-    thread.daemon = True
-    thread.start()
-    called = True
+    self._mgr = OpenRTM_aist.Manager.instance()
+    self._mgr.addManagerActionListener(ManagerActionListener(self))
 
     return
 
@@ -387,9 +355,27 @@ class InPortPahoSubscriber(OpenRTM_aist.InPortProvider, PahoSubscriber):
 
 
 ##
-# @brief Catch ctrl+c interruption
+# @class ManagerActionListener
+# @brief ManagerActionListener class
 #
-signal.signal(signal.SIGINT, InPortPahoSubscriber.signal_handler)
+class ManagerActionListener(ManagerActionListeners):
+  def __init__(self, InPortPahoSubscriber):
+    self._InPortPahoSubscriber = InPortPahoSubscriber
+
+  def preShutdown(self):
+    pass
+
+  ##
+  # @brief Clean up mqtt communication module instance when RTC exit
+  #
+  def postShutdown(self):
+    self._InPortPahoSubscriber.__del__()
+
+  def preReinit(self):
+    pass
+
+  def postReinit(self):
+    pass
 
 ##
 # @brief Initialize InPortPahoSubscriber module